Update innovation.css
I've had a lot of trouble getting either the client or the command-line of git to work. I've given up, and decided to just update it via the browser. I added the scss files in the sass folder, and will update the server via filezilla, but feel free to push github over it.